+++ vcs/src/main/java/vcs/commands/impl/CheckoutCommand.java
@@ -1,7 +1,8 @@
-package vcs.commands;
+package vcs.commands.impl;
 
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
+import vcs.commands.Command;
 import vcs.config.CommitConfig;
 import vcs.config.GlobalConfig;
 import vcs.graph.Commit;
