+++ vcs/src/main/java/vcs/commands/impl/InitCommand.java
@@ -1,6 +1,7 @@
-package vcs.commands;
+package vcs.commands.impl;
 
 import com.beust.jcommander.Parameters;
+import vcs.commands.Command;
 import vcs.config.CommitConfig;
 import vcs.config.GlobalConfig;
 import vcs.util.VcsUtils;
